Moved the DetailsView and IconsView to new files, because convention says one class per file.
